+++ abort-controller.d.ts
@@ -1,6 +1,7 @@
+import { type Duration } from './duration.js';
 export declare class TimeoutAbortController extends AbortController {
     constructor(options: {
         signal?: AbortSignal;
-        timeout: number;
+        timeout: Duration;
     });
 }