+++ duration.d.ts
@@ -1,0 +1,2 @@
+export type Duration = number | `${number}s` | `${number}ms`;
+export declare function toMs(duration: Duration): number;