+++ index.js
@@ -1,8 +1,8 @@
 export * from './abort-controller.js';
+export * from './duration.js';
 export * from './error.js';
 export * from './misc.js';
 export * from './promise.js';
-export * from './request.js';
+export * from './fetch.js';
 export * from './signature.js';
 export * from './type.js';
-//# sourceMappingURL=index.js.map