+++ matcher/jest.d.ts
@@ -1,0 +1,12 @@
+import type { CustomMatchers } from './core.js';
+export * from './core.js';
+declare global {
+    namespace jest {
+        interface Expect extends CustomMatchers {
+        }
+        interface Matchers<R> extends CustomMatchers<R> {
+        }
+        interface InverseAsymmetricMatchers extends CustomMatchers {
+        }
+    }
+}