+++ request.d.ts
@@ -1,3 +1,0 @@
-export declare function request(url: string, options?: RequestInit & {
-    timeout?: number;
-}): Promise<Response>;