+++ type.d.ts
@@ -1,9 +1,3 @@
-declare global {
-    interface BigInt {
-        toJSON: () => string;
-    }
-}
-export declare const NIL: undefined;
 export declare const Nil: undefined;
 export type Nil = typeof undefined;
 export type NumStr = `${number}`;
@@ -15,7 +9,7 @@ export type BigIntOrNil = TypeOrNil<bigint>;
 export type DateOrNil = TypeOrNil<Date>;
 export type NumberOrNil = TypeOrNil<number>;
 export type StringOrNil = TypeOrNil<string>;
-export type BufferOrNil = TypeOrNil<Buffer>;
+export type Uint8ArrayOrNil = TypeOrNil<Uint8Array>;
 export type NumStrOrNil = TypeOrNil<NumStr>;
 export declare function toBigInt(x?: unknown): bigint;
 export declare function toBigIntOrNil(x?: unknown): BigIntOrNil;